🚀 refactor(lotto): Map에 모든 결과를 저장한 후에 리턴하며 LottoResults 객체를 생성하도록 변경
